chore(numpy2): migrate deprecated dtype aliases across core modules
Replace np.complex_, np.float_, np.int_, np.bool_ with canonical equivalents
(np.complex128, np.float64, int, np.bool_) required for NumPy 2.4 compatibility.
Fixes Gemini regression in ResultsIO.py where np.bool_ was incorrectly replaced
with bool, silently breaking isinstance checks for numpy boolean scalars in the
MATLAB export path. RefProfile cache files reverted to keep existing numerical
baseline; re-baseline intentionally when EOS changes are made deliberately.

+++ PlanetProfile/MagneticInduction/MoonMag/trajec_analysis.py
@@ -192,8 +192,8 @@ def fitData(bname, recalcMoments=True, recalcData=True, recalcFlybys=True, do_pa
             _, nvals[j], mvals[j], lin_Binm_Re, lin_Binm_Im = np.loadtxt(BinmFiles[j], skiprows=1, unpack=True,
                                                                          delimiter=',')
             Binm[j] = np.reshape(lin_Binm_Re + 1j * lin_Binm_Im, (np.size(peak_omegas), -1))
-            nvals[j] = nvals[j].astype(np.int_)
-            mvals[j] = mvals[j].astype(np.int_)
+            nvals[j] = nvals[j].astype(int)
+            mvals[j] = mvals[j].astype(int)
 
         for i, thisFlyby in np.ndenumerate(fbList):
             print(f"Getting square differences for {thisFlyby.upper()}, #{i[0]+1} of {nFlybys}.")
